Conflicts:
	common/mekanism/client/ModelRendererSelectiveFace.java

+++ common/mekanism/client/ModelRendererSelectiveFace.java
@@ -21,22 +21,30 @@ public class ModelRendererSelectiveFace
 {
     public float textureWidth;
     public float textureHeight;
+    
     public float offsetX;
     public float offsetY;
     public float offsetZ;
+    
     public float rotationPointX;
     public float rotationPointY;
     public float rotationPointZ;
+    
     public float rotateAngleX;
     public float rotateAngleY;
     public float rotateAngleZ;
+    
     public boolean mirror;
     public boolean showModel;
     public boolean isHidden;
+    
     public List<ModelBoxSelectiveFace> cubeList = new ArrayList<ModelBoxSelectiveFace>();
+    
     private int textureOffsetX;
     private int textureOffsetY;
+    
     private ModelBase baseModel;
+    
     private Map<BooleanArray, DisplayInteger> displayLists = new HashMap<BooleanArray, DisplayInteger>();
     
 
@@ -46,6 +54,7 @@ public class ModelRendererSelectiveFace
         textureHeight = 32.0F;
         showModel = true;
         baseModel = modelBase;
+        
         setTextureSize(modelBase.textureWidth, modelBase.textureHeight);
     }
 
@@ -59,6 +68,7 @@ public class ModelRendererSelectiveFace
     {
         textureOffsetX = offsetX;
         textureOffsetY = offsetY;
+        
         return this;
     }
 
@@ -78,27 +88,28 @@ public class ModelRendererSelectiveFace
     @SideOnly(Side.CLIENT)
     public void render(boolean[] dontRender, float scaleFactor)
     {
-        if (!isHidden)
+        if(!isHidden)
         {
-            if (showModel)
+            if(showModel)
             {
             	DisplayInteger currentDisplayList = displayLists.get(new BooleanArray(dontRender));
-                if (currentDisplayList == null)
+            	
+                if(currentDisplayList == null)
                 {
                     currentDisplayList = compileDisplayList(dontRender, scaleFactor);
                 }
 
                 GL11.glTranslatef(offsetX, offsetY, offsetZ);
+                
                 int i;
 
-                if (rotateAngleX == 0.0F && rotateAngleY == 0.0F && rotateAngleZ == 0.0F)
+                if(rotateAngleX == 0.0F && rotateAngleY == 0.0F && rotateAngleZ == 0.0F)
                 {
-                    if (rotationPointX == 0.0F && rotationPointY == 0.0F && rotationPointZ == 0.0F)
+                    if(rotationPointX == 0.0F && rotationPointY == 0.0F && rotationPointZ == 0.0F)
                     {
                         currentDisplayList.render();
                     }
-                    else
-                    {
+                    else {
                         GL11.glTranslatef(rotationPointX * scaleFactor, rotationPointY * scaleFactor, rotationPointZ * scaleFactor);
                         currentDisplayList.render();
                         GL11.glTranslatef(-rotationPointX * scaleFactor, -rotationPointY * scaleFactor, -rotationPointZ * scaleFactor);
@@ -109,17 +120,17 @@ public class ModelRendererSelectiveFace
                     GL11.glPushMatrix();
                     GL11.glTranslatef(rotationPointX * scaleFactor, rotationPointY * scaleFactor, rotationPointZ * scaleFactor);
 
-                    if (rotateAngleZ != 0.0F)
+                    if(rotateAngleZ != 0.0F)
                     {
                         GL11.glRotatef(rotateAngleZ * (180F / (float)Math.PI), 0.0F, 0.0F, 1.0F);
                     }
 
-                    if (rotateAngleY != 0.0F)
+                    if(rotateAngleY != 0.0F)
                     {
                         GL11.glRotatef(rotateAngleY * (180F / (float)Math.PI), 0.0F, 1.0F, 0.0F);
                     }
 
-                    if (rotateAngleX != 0.0F)
+                    if(rotateAngleX != 0.0F)
                     {
                         GL11.glRotatef(rotateAngleX * (180F / (float)Math.PI), 1.0F, 0.0F, 0.0F);
                     }
@@ -139,20 +150,22 @@ public class ModelRendererSelectiveFace
         DisplayInteger displayList = DisplayInteger.createAndStart();
         Tessellator tessellator = Tessellator.instance;
 
-        for (int i = 0; i < cubeList.size(); ++i)
+        for(int i = 0; i < cubeList.size(); ++i)
         {
             cubeList.get(i).render(tessellator, dontRender, scaleFactor);
         }
 
         displayList.endList();
         displayLists.put(new BooleanArray(dontRender), displayList);
+        
         return displayList;
     }
 
     public ModelRendererSelectiveFace setTextureSize(int sizeX, int sizeY)
     {
-        textureWidth = (float)sizeX;
-        textureHeight = (float)sizeY;
+        textureWidth = sizeX;
+        textureHeight = sizeY;
+        
         return this;
     }
 }
